Edits so far
lmk if you guys have any suggestions. Not done, I intend to make the card and the elements inside not expand past a certain amount so it doesnt throw off the items on screen, as well as make it where you can color your tasks. also  any ideas on how to display task (yellow square)?